requirements.txt
razdel
pandas
numpy
scikit-learn
bertopic
sentence-transformers
umap-learn
hdbscan
nltk
matplotlib
seaborn
plotly
peft
bitsandbytes
accelerate
torch
torchvision
torchaudio 
bertopic>=0.15.0
sentence-transformers>=2.4.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
umap-learn>=0.5.0
hdbscan>=0.8.0
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
transformers>=4.21.0
torch>=1.12.0
pathlib
pickle-mixin 